+++ virspit-order/src/main/java/com/virspit/virspitorder/service/NftService.java
@@ -1,0 +1,5 @@
+package com.virspit.virspitorder.service;
+
+public class NftService {
+
+}
