+++ virspit-order/src/main/java/com/virspit/virspitorder/service/OrderService.java
@@ -1,8 +1,8 @@
 package com.virspit.virspitorder.service;
 
 import com.virspit.virspitorder.dto.response.OrdersResponseDto;
-import com.virspit.virspitorder.common.error.ErrorCode;
-import com.virspit.virspitorder.common.error.exception.BusinessException;
+import com.virspit.virspitorder.response.error.ErrorCode;
+import com.virspit.virspitorder.response.error.exception.BusinessException;
 import com.virspit.virspitorder.repository.OrderRepository;
 import com.virspit.virspitorder.util.StringUtils;
 import lombok.RequiredArgsConstructor;
